requirements.txt
# Core Django
Django==6.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.4.1
django-cors-headers==4.6.0

# Environment variables
python-dotenv==1.0.1

# Production server
gunicorn==22.0.0
whitenoise==6.7.0

# Database (PostgreSQL for production)
psycopg2-binary==2.9.9

# HTTP requests
requests==2.31.0

# Celery for async tasks
celery==5.4.0
redis==5.1.1

# Groq AI API
groq==0.11.0
